feat: resto

+++ start.sh
@@ -5,40 +5,50 @@ SRC="$1"
 OUT="/tmp/programaPaia.o"
 
 rm -f "$OUT" > /dev/null 2>&1
-clang -std=c99 -O2 -lm "${SRC}/code.c" -o  "$OUT"
+clang -std=c99 -O2 -lm "${SRC}/code.c" -o "$OUT"
 
 set +e
- "$OUT" < "${SRC}/input.txt"
-set -e
 
+# Itera sobre todos os arquivos input*
+for IN_FILE in "${SRC}"/input*; do
+    [ ! -f "$IN_FILE" ] && continue
 
-mapfile -t OUT_LINES < <("$OUT" < "${SRC}/input.txt")
-mapfile -t EXP_LINES < "${SRC}/result.txt"
+    # Define o arquivo de resultado correspondente
+    BASE=$(basename "$IN_FILE")
+    RES_FILE="${SRC}/result${BASE#input}"
 
-rm -f "$OUT" > /dev/null 2>&1
+    [ ! -f "$RES_FILE" ] && continue
 
-echo "====================================================================="
-printf "| %5s | %-30s | %-30s | %-8s |\n" "LINHA" "ESPERADO" "SAÍDA" "STATUS"
-echo "====================================================================="
+    mapfile -t OUT_LINES < <("$OUT" < "$IN_FILE")
+    mapfile -t EXP_LINES < "$RES_FILE"
 
-max=${#OUT_LINES[@]}
-[ ${#EXP_LINES[@]} -gt $max ] && max=${#EXP_LINES[@]}
+    echo ""
+    echo "=================== Comparando $IN_FILE ========================================"
+    printf "| %5s | %-30s | %-30s | %-8s |\n" "LINHA" "ESPERADO" "SAÍDA" "STATUS"
+    echo "====================================================================================================="
 
-truncate() {
-    local s="$1"
-    local n="${2:-30}"
-    [ "${#s}" -le "$n" ] && printf "%s" "$s" || printf "%s..." "${s:0:((n-3))}"
-}
+    max=${#OUT_LINES[@]}
+    [ ${#EXP_LINES[@]} -gt $max ] && max=${#EXP_LINES[@]}
 
-for ((i=0; i<max; i++)); do
-    e="${EXP_LINES[i]:-}"
-    o="${OUT_LINES[i]:-}"
+    truncate() {
+        local s="$1"
+        local n="${2:-30}"
+        [ "${#s}" -le "$n" ] && printf "%s" "$s" || printf "%s..." "${s:0:((n-3))}"
+    }
 
-    st="FAILED"
-    [ "$o" = "$e" ] && st="PASS"
+    for ((i=0; i<max; i++)); do
+        e="${EXP_LINES[i]:-}"
+        o="${OUT_LINES[i]:-}"
 
-    printf "| %5d | %-30s | %-30s | %-8s |\n" \
-        "$((i+1))" "$(truncate "$e")" "$(truncate "$o")" "$st"
+        st="FAILED"
+        [ "$o" = "$e" ] && st="PASS"
+
+        printf "| %5d | %-30s | %-30s | %-8s |\n" \
+            "$((i+1))" "$(truncate "$e")" "$(truncate "$o")" "$st"
+    done
+
+    echo "====================================================================================================="
+    echo ""
 done
 
-echo "====================================================================="
+rm -f "$OUT" > /dev/null 2>&1
